Updated Font Data
Woops! Previous push did not have all of the correct files. This branch has the data & png files for 'Kawaii' & 'KawaiiCompressed'

Kawaii is good for small user interfaces; it is readable.

KawaiiCompressed is good for using the font as a tileset, or for visual fx in roguelikes with particle systems.