Kawaii font addition (#321)
* Added font, 'Kawaii' to Fonts folder

I made this custom font by myself. each character is 6x3 pixels. It is a great font for visual fx / npc names in roguelikes.

* Updated Font Data

Woops! Previous push did not have all of the correct files. This branch has the data & png files for 'Kawaii' & 'KawaiiCompressed'

Kawaii is good for small user interfaces; it is readable.

KawaiiCompressed is good for using the font as a tileset, or for visual fx in roguelikes with particle systems.